add types and folder in header component
