+++ src/components/modules/Header/Header.types.ts
@@ -1,0 +1,14 @@
+interface NavigationTypes {
+    name: string;
+    link: string;
+}
+interface NavigationSubItem {
+    icon: string;
+}
+
+interface NavigationItem {
+    subList?: Array<NavigationTypes & NavigationSubItem>;
+    moreLink?: NavigationTypes;
+}
+
+export type interfaceNavigationItem = NavigationTypes & NavigationItem;
